+++ SandboxiePlus/QSbieAPI/SbieStatus.h
@@ -125,13 +125,10 @@ public:
 	CSbieProgress() : m_Status(OP_ASYNC), m_Canceled(false) {}
 
 	void Cancel() { m_Canceled = true; }
-	bool IsCancel() { return m_Canceled; }
+	bool IsCanceled() { return m_Canceled; }
 
-	void ShowMessage(const QString& text) { emit Message(text);
-#ifdef _DEBUG
-		//QThread::msleep(500);
-#endif
-	}
+	void ShowMessage(const QString& text) { emit Message(text);}
+	void SetProgress(int value) { emit Progress(value); }
 	void Finish(SB_STATUS status) { m_Status = m_Canceled ? SB_ERR(OP_CANCELED) : status; emit Finished(); }
 
 	SB_STATUS GetStatus() { return m_Status; }
@@ -140,6 +137,7 @@ public:
 signals:
 	//void Progress(int procent);
 	void Message(const QString& text);
+	void Progress(int value);
 	void Finished();
 
 protected: